Development PR

+++ icons/build-favicons.sh
@@ -2,6 +2,7 @@
 set -euf
 
 mode=''
+only_pattern=''
 if [ "${HOMELAB_ENV-}" != '' ]; then
     mode="$HOMELAB_ENV"
 fi
@@ -15,6 +16,10 @@ while [ "$#" -gt 0 ]; do
         mode='prod'
         shift
         ;;
+    --only)
+        only_pattern="$2"
+        shift 2
+        ;;
     *)
         printf 'Unknown argument %s\n' "$1"
         exit 1
@@ -25,7 +30,9 @@ HOMELAB_ENV="$mode"
 
 input_dir="$(git rev-parse --show-toplevel)/icons"
 output_dir="$(git rev-parse --show-toplevel)/docker-images/custom/favicons/icons"
-rm -rf "$output_dir"
+if [ "${only_pattern:-}" = '' ]; then
+    rm -rf "$output_dir"
+fi
 mkdir -p "$output_dir"
 
 tmpdir=''
@@ -37,6 +44,10 @@ default_image_size='64x64'
 # shellcheck disable=SC2034
 default_convert_options='magick -density 2000 -background none -bordercolor transparent INPUT_FILE -resize RESOLUTION -density 2000 OUTPUT_FILE'
 
+## Dashboard Icons ##
+
+convert_image_full "$input_dir/gitman-repositories/dashboard-icons/svg/prometheus-node-exporter.svg" "$output_dir/nodeexporter.png"
+
 # Smtp4dev
 
 convert_image_draft 'magick -background none INPUT_FILE -resize 16x16 -density 1200 OUTPUT_FILE' "$input_dir/other/smtp4dev-favicon.png" "$tmpdir/smtp4dev-favicon-16.png"
@@ -45,11 +56,11 @@ convert_image_draft 'magick -background none INPUT_FILE -resize 32x32 -density 1
 optimize_image "$tmpdir/smtp4dev-favicon-32.png"
 convert_ico "$tmpdir/smtp4dev-favicon-16.png $tmpdir/smtp4dev-favicon-32.png" "$(git rev-parse --show-toplevel)/docker-images/external/smtp4dev/icons/favicon.ico"
 convert_image_full "$input_dir/other/smtp4dev-favicon.png" "$(git rev-parse --show-toplevel)/docker-images/external/smtp4dev/icons/favicon.png"
+rm -f "$tmpdir/smtp4dev-favicon-16.png" "$tmpdir/smtp4dev-favicon-32.png"
 
 ## Other ##
 
 convert_image_full "$input_dir/gitman-repositories/homer-icons/png/docker-moby.png" "$output_dir/docker-stats.png"
-convert_image_full "$input_dir/gitman-repositories/kubernetes-community/icons/svg/resources/unlabeled/ns.svg" "$output_dir/nodeexporter.png"
 convert_image_full "$input_dir/other/renovatebot.png" "$output_dir/renovatebot.png"
 convert_image_full "$input_dir/gitman-repositories/dashboard-icons/svg/samba-server.svg" "$output_dir/samba.png"
 convert_image_full "$input_dir/gitman-repositories/dashboard-icons/svg/unbound.svg" "$output_dir/unbound.png"
